requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
mlflow==2.13.0
streamlit==1.37.0
pyyaml==6.0.1
joblib==1.4.2
plotly==5.22.0
flake8==7.0.0
pytest==8.2.0
boto3==1.34.162
